feat: optimize file structure
