--- packages/vmind/src/application/base/index.ts
@@ -1,1 +1,0 @@
-export class BaseApplication {}
