--- packages/vmind/src/application/base/types.ts
@@ -1,3 +1,0 @@
-export interface IApplication {
-  tasks: TaskNode[];
-}
