+++ packages/vmind/src/base/tools/chatManager/index.ts
