+++ packages/vmind/src/base/tools/chatManager/types.ts
