+++ packages/vmind/src/base/tools/patcher/types.ts
