+++ packages/vmind/src/base/tools/prompt/index.ts
