+++ packages/vmind/src/base/tools/prompt/types.ts
