+++ packages/vmind/src/gpt/dataProcess/query/queryDataset.ts
@@ -1,7 +1,7 @@
 import { DataItem, ILLMOptions, SimpleFieldInfo } from '../../../typings';
 import { parseGPTQueryResponse, parseRespondField, patchQueryInput } from './utils';
 import { DataQueryResponse } from './type';
-import { parseGPTResponse as parseGPTResponseAsJSON, requestGPT } from '../../../taskNode/utils';
+import { parseGPTResponse as parseGPTResponseAsJSON, requestGPT } from '../../../base/taskNode/utils';
 import { getQueryDatasetPrompt } from '../prompts';
 import { queryDataset } from '../../../common/dataProcess/dataQuery';
