--- packages/vmind/src/types/index.ts
@@ -1,1 +1,0 @@
-export type Context = Record<string, string>;
